+++ db/downgrade_0058---0057.sql
@@ -1,0 +1,5 @@
+BEGIN;
+
+ALTER TABLE body DROP blank_updates_permitted;
+
+COMMIT;