+++ db/schema.sql
@@ -54,6 +54,8 @@ create table body (
     suppress_alerts boolean not null default 'f',
     can_be_devolved boolean not null default 'f',
     send_extended_statuses boolean not null default 'f',
+    fetch_problems boolean not null default 'f',
+    blank_updates_permitted boolean not null default 'f',
     deleted boolean not null default 'f'
 );