+++ db/schema_0057-fetch-problems.sql
@@ -1,0 +1,5 @@
+BEGIN;
+
+ALTER TABLE body ADD fetch_problems boolean default 'f' not null;
+
+COMMIT;