+++ perllib/FixMyStreet/App/Controller/Admin.pm
@@ -445,6 +445,8 @@ sub body_params : Private {
     my %defaults = map { $_ => '' } @fields;
     %defaults = ( %defaults,
         send_comments => 0,
+        fetch_problems => 0,
+        blank_updates_permitted => 0,
         suppress_alerts => 0,
         comment_user_id => undef,
         send_extended_statuses => 0,